Remove PaymentLottieView and ConfirmPayment from SendBitcoin
