+++ app/screens/send-bitcoin-screen/error_lottie.json
