+++ app/screens/send-bitcoin-screen/index.ts
@@ -1,2 +1,3 @@
 export * from "./send-bitcoin-screen"
 export * from "./scanning-qrcode-screen"
+export * from "./send-bitcoin-confirmation-screen"
