+++ app/screens/send-bitcoin-screen/pending_lottie.json
