+++ app/screens/send-bitcoin-screen/success_lottie.json
